requirements.txt
-f https://download.pytorch.org/whl/cu111/torch_stable.html

autopep8==1.5.7
datasets==1.11.0
torch==1.9.0+cu111
torchvision==0.10.0
pytorch-lightning==1.5.8
coloredlogs==15.0.1
scikit-image==0.17.2
timm==0.4.12
Pillow==8.3.1
transformers==4.10.2
scikit-image==0.17.2
scikit-learn==0.24.2
dataclasses
pycocotools==2.0.2
seqeval==1.2.2
torchtext==0.10.0
dacite==1.6.0
datadings==3.4.2
setuptools==59.5.0
opencv-python